create helpers for base, more_info, and search
